update mapping object on weeklyplanitem

+++ Com.Bateeq.Service.Masterplan.Lib/ViewModels/WeeklyPlanItemViewModel.cs
@@ -11,10 +11,10 @@ namespace Com.Bateeq.Service.Masterplan.Lib.ViewModels
         public int Efficiency { get; set; }
         public int Operator { get; set; }
         public int WorkingHours { get; set; }
-        public int AhTotal { get; set; }
-        public int EhTotal { get; set; }
-        public int UsedEh { get; set; }
-        public int RemainingEh { get; set; }
+        public double AhTotal { get; set; }
+        public double EhTotal { get; set; }
+        public double UsedEh { get; set; }
+        public double RemainingEh { get; set; }
         public DateTimeOffset StartDate { get; set; }
         public DateTimeOffset EndDate { get; set; }
     }
